+++ src/components/shared.styled.tsx
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+import { Flex } from "@sanity/ui";
+
+export const GiphyUserImage = styled.img`
+  height: 32px;
+  width: 32px;
+  margin: 0 0.5rem;
+`;
+
+export const Spacer = styled.div`
+  flex-grow: 1;
+`;
+
+export const Flex100Center = styled(Flex)`
+  width: 100%;
+`;