+++ src/types/index.ts
@@ -1,7 +1,37 @@
-type GiphyImageItem = {
+export type GiphyImageItem = {
   url: string;
   mp4?: string;
+  height: number;
+  width: number;
+  size?: number;
 };
+
+export interface GiphyUser {
+  id: number;
+  avatar_url: string;
+  banner_image: string;
+  banner_url: string;
+  profile_url: string;
+  username: string;
+  display_name: string;
+  name: string;
+  attribution_display_name: string;
+  description: string;
+  user_type: string;
+  facebook_url: string;
+  instagram_url: string;
+  twitter_url: string;
+  twitter: string;
+  tumblr_url: string;
+  website_url: string;
+  website_display_url: string;
+  is_public: boolean;
+  is_staff: boolean;
+  is_superuser: boolean;
+  is_verified: boolean;
+  suppress_chrome: boolean;
+}
+
 export interface GiphyResult {
   type: string;
   id: string;
@@ -13,7 +43,7 @@ export interface GiphyResult {
   username: string;
   source: string;
   title: string;
-  rating: string;
+  rating: GiphyRating;
   content_url: string;
   source_tld: string;
   source_post_url: string;
@@ -21,30 +51,41 @@ export interface GiphyResult {
   import_datetime: string;
   trending_datetime: string;
   images: Record<ImageTypes, GiphyImageItem>;
-  user: {
-    avatar_url: string;
-    banner_image: string;
-    banner_url: string;
-    profile_url: string;
-    username: string;
-    display_name: string;
-    description: string;
-    instagram_url: string;
-    website_url: string;
-    is_verified: true;
-  };
-  analytics_response_payload: string;
-  analytics: {
-    onload: {
-      url: string;
-    };
-    onclick: {
-      url: string;
-    };
-    onsent: {
-      url: string;
-    };
-  };
+  user: GiphyUser;
+}
+
+export interface RelatedGif {
+  type: "gif";
+  id: string;
+  index_id: number;
+  url: string;
+  slug: string;
+  bitly_gif_url: string;
+  bitly_url: string;
+  embed_url: string;
+  username: string;
+  source: string;
+  title: string;
+  rating: GiphyRating;
+  content_url: string;
+  tags: string[];
+  featured_tags: string[];
+  user_tags: [];
+  source_tld: string;
+  source_post_url: string;
+  is_hidden: boolean;
+  is_removed: boolean;
+  is_community: boolean;
+  is_anonymous: boolean;
+  is_featured: boolean;
+  is_realtime: boolean;
+  is_sticker: boolean;
+  import_datetime: string;
+  trending_datetime: string;
+  create_datetime: string;
+  update_datetime: string;
+  images: Record<string, GiphyImageItem>;
+  user?: GiphyUser;
 }
 
 export enum ImageTypes {
@@ -69,4 +110,35 @@ export enum ImageTypes {
   FixedWidthSmall = "fixed_width_small",
   FixedWidthSmallStill = "fixed_width_small_still",
   FixedWidthStill = "fixed_width_still",
+}
+
+export const SupportedImageTypes: ImageTypes[] = [
+  ImageTypes.Original,
+  ImageTypes.Downsized,
+  ImageTypes.FixedHeight,
+  ImageTypes.FixedWidth,
+  ImageTypes.OriginialStill,
+];
+
+export enum SearchTypes {
+  Search = "search",
+  Trending = "trending",
+  Random = "random",
+  Related = "related",
+}
+
+export enum GiphyRating {
+  G = "g",
+  PG = "pg",
+  "PG-13" = "pg-13",
+  R = "r",
+}
+
+export enum Tone {
+  positive = "positive",
+  primary = "primary",
+  caution = "caution",
+  critical = "critical",
+  brand = "brand",
+  default = "default",
 }