+++ CHANGELOG.md
@@ -1,9 +1,57 @@
-## [3.0.1](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v3.0.0...v3.0.1) (2022-06-18)
+## [4.0.3](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v4.0.2...v4.0.3) (2022-06-27)
+
+
+### Bug Fixes
+
+* Expose search term in the provider. Remove unused props and logs ([76ef50c](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/76ef50cd5a5b8b0032ab01b53546701c55959e1c))
+
+## [4.0.2](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v4.0.1...v4.0.2) (2022-06-26)
+
+
+### Bug Fixes
+
+* allow to open current asset without closing giphy dialog ([9886b0a](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/9886b0ae7a0cce35821d6e59a86c3252e95cdeb8))
+* centering done ([b5f2272](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/b5f2272fa403255112d1c3023a994e70306840f2))
+* restructure to allow dialog to open dialog from anywhere ([6e1b032](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/6e1b032827e99eff9c61bc1883e121d37440a115))
+* style fixes. centering ([894b9bf](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/894b9bf8a531491bc4ba76fba680773d8466d1d5))
+
+## [4.0.1](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v4.0.0...v4.0.1) (2022-06-23)
+
+
+### Bug Fixes
+
+* lib inside src should not be ignored ([520e474](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/520e4740949b2f08e31b9b00ea0964c820a14ccc))
+
+# [4.0.0](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v3.1.0...v4.0.0) (2022-06-23)
+
+
+### Bug Fixes
+
+* fix ref to container removed before unobserve ([99af551](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/99af551189411046f988f52288d3726605180dca))
+
+
+### Features
+
+* remove should auto play config ([abca7b7](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/abca7b7320c1f080581a25820904ce6ee96c1549))
+
+
+### BREAKING CHANGES
+
+* removing it changes how this works. no deprecation grace period either
+
+# [3.1.0](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v3.0.0...v3.1.0) (2022-06-22)
 
 
 ### Bug Fixes
 
 * force semantic-release ([f4c7c23](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/f4c7c2376d3a0899af4806570c5c2cad3b4831c3))
+
+
+### Features
+
+* add masonry ([7ff1bdc](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/7ff1bdcc279e0c09d70f889c86077d76d34d57fe))
+* reduce options ([23d2772](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/23d2772c17e821cf9d67a1ba2652a4d9d1effb49))
+* refactor and introduce related gifs. dialog ([9f84f37](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/commit/9f84f37dda967ebcd0ac3861922716ddb172a4bb))
 
 # [3.0.0](https://github.com/XSlemX/sanity-plugin-asset-source-giphy/compare/v2.0.0...v3.0.0) (2022-06-18)